Added Lighthouse CI configuration and updated Github Actions workflow
A new file, lighthouserc.js is added to specify settings for Lighthouse CI. Changes are made in Github Actions workflow file (.github/workflows/lighthouse-ci.yml) to adhere to the same and run Lighthouse CI every time a push event occurs. The purpose of this is to maintain high-quality code and identify potential issues earlier in the development process.